Update Bazel artifacts with new build strategy.
Bazel has introduced Bazelisk, which provides a better strategy for
keeping Bazel versions across different developer machines in sync.

This commit also updates the canonical project view file with a list of
run configurations and points to Bazelisk as the Bazel binary.

Finally, add a .bazelversion file so that we have more control over when
we upgrade Bazel versions across all development machines.